Removed loading states since we dont use them
