+++ Projects/EditCoInsured/Sources/Stores/EditCoInsuredState.swift
@@ -6,6 +6,5 @@ import hCore
 import hGraphQL
 
 public struct EditCoInsuredState: StateProtocol {
-    var loadingStates: [EditCoInsuredAction: LoadingState<String>] = [:]
     public init() {}
 }
